Fix #18: Implementatie van de flow state transities voor een woordenboek artikel.  (#25)
