--- app/Filament/Resources/WordResource/Pages/EditWord.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Filament\Resources\WordResource\Pages;
-
-use App\Filament\Resources\WordResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-use Filament\Resources\Pages\EditRecord\Concerns\HasWizard;
-
-final class EditWord extends EditRecord
-{
-    protected static string $resource = WordResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\DeleteAction::make()
-                ->icon('heroicon-o-trash'),
-        ];
-    }
-}
