--- app/Filament/Resources/WordResource/Pages/ListWords.php
@@ -1,22 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Filament\Resources\WordResource\Pages;
-
-use App\Filament\Resources\WordResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-final class ListWords extends ListRecords
-{
-    protected static string $resource = WordResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make()
-                ->icon('heroicon-o-plus'),
-        ];
-    }
-}
