--- app/Filament/Resources/WordResource/Pages/ViewWord.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Filament\Resources\WordResource\Pages;
-
-use App\Filament\Resources\WordResource;
-use Filament\Actions\DeleteAction;
-use Filament\Actions\EditAction;
-use Filament\Resources\Pages\ViewRecord;
-
-final class ViewWord extends ViewRecord
-{
-    protected static string $resource = WordResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            EditAction::make()->icon('heroicon-o-pencil-square'),
-            DeleteAction::make()->icon('heroicon-o-trash'),
-        ];
-    }
-}
